HAHAHAHAHAHAHAHAHAHAH

+++ src/main/java/engsoft/ProgressaoGeometrica.java
@@ -8,7 +8,7 @@ public class ProgressaoGeometrica extends Progressao {
 	this.progression = new FirstOrderProgression<Double>(new Geometric(2.0), 1.0);
     }
 
-    public ProgressaoGeometrica(double base) {
-	this.progression = new FirstOrderProgression<Double>(new Geometric(base), 1.0);
+    public ProgressaoGeometrica(int base) {
+	this.progression = new FirstOrderProgression<Double>(new Geometric((double) base), 1.0);
     }
 }
